refactor(operator): extract MembershipsTab from SolverNetsSection JOINED block
Moves the joined SolverNets list + query into MembershipsTab (§2.4);
SolverNetsSection is stripped to DISCOVER block only (still used in
Operator.tsx, deleted in next commit). Hash-deep-link logic dropped
(proper sub-routes replace it). 3 new tests covering count header,
empty state, and card rendering.
